backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-multipart==0.0.20
chromadb==0.5.23
PyYAML==6.0.2
requests==2.32.3
posthog==3.5.0
Markdown==3.7
python-docx==1.1.2
